refactor: refactor context proxy
